requirements.txt
pytorch-lightning
torch
numpy
tensorboard
pillow
torchvision